+++ src/ray/rpc/gcs_server/gcs_rpc_server.h
@@ -1,5 +1,4 @@
-#ifndef RAY_RPC_GCS_RPC_SERVER_H
-#define RAY_RPC_GCS_RPC_SERVER_H
+#pragma once
 
 #include "src/ray/rpc/grpc_server.h"
 #include "src/ray/rpc/server_call.h"
@@ -9,9 +8,9 @@
 namespace ray {
 namespace rpc {
 
-class JobInfoHandler {
+class JobInfoAccessHandler {
  public:
-  virtual ~JobInfoHandler() = default;
+  virtual ~JobInfoAccessHandler() = default;
 
   virtual void HandleAddJob(const AddJobRequest &request, AddJobReply *reply,
                             SendReplyCallback send_reply_callback) = 0;
@@ -21,14 +20,14 @@ class JobInfoHandler {
                                      SendReplyCallback send_reply_callback) = 0;
 };
 
-/// The `GrpcService` for `JobInfoGcsService`.
-class JobInfoGrpcService : public GrpcService {
+/// The `GrpcService` for `JobInfoAccessService`.
+class JobInfoAccessGrpcService : public GrpcService {
  public:
   /// Constructor.
   ///
   /// \param[in] handler The service handler that actually handle the requests.
-  explicit JobInfoGrpcService(boost::asio::io_service &io_service,
-                              JobInfoHandler &handler)
+  explicit JobInfoAccessGrpcService(boost::asio::io_service &io_service,
+                                    JobInfoAccessHandler &handler)
       : GrpcService(io_service), service_handler_(handler){};
 
  protected:
@@ -39,30 +38,29 @@ class JobInfoGrpcService : public GrpcService {
       std::vector<std::pair<std::unique_ptr<ServerCallFactory>, int>>
           *server_call_factories_and_concurrencies) override {
     std::unique_ptr<ServerCallFactory> add_job_call_factory(
-        new ServerCallFactoryImpl<JobInfoGcsService, JobInfoHandler, AddJobRequest,
-                                  AddJobReply>(
-            service_, &JobInfoGcsService::AsyncService::RequestAddJob, service_handler_,
-            &JobInfoHandler::HandleAddJob, cq, main_service_));
+        new ServerCallFactoryImpl<JobInfoAccessService, JobInfoAccessHandler,
+                                  AddJobRequest, AddJobReply>(
+            service_, &JobInfoAccessService::AsyncService::RequestAddJob,
+            service_handler_, &JobInfoAccessHandler::HandleAddJob, cq, main_service_));
     server_call_factories_and_concurrencies->emplace_back(std::move(add_job_call_factory),
                                                           1);
 
     std::unique_ptr<ServerCallFactory> mark_job_finished_call_factory(
-        new ServerCallFactoryImpl<JobInfoGcsService, JobInfoHandler,
+        new ServerCallFactoryImpl<JobInfoAccessService, JobInfoAccessHandler,
                                   MarkJobFinishedRequest, MarkJobFinishedReply>(
-            service_, &JobInfoGcsService::AsyncService::RequestMarkJobFinished,
-            service_handler_, &JobInfoHandler::HandleMarkJobFinished, cq, main_service_));
+            service_, &JobInfoAccessService::AsyncService::RequestMarkJobFinished,
+            service_handler_, &JobInfoAccessHandler::HandleMarkJobFinished, cq,
+            main_service_));
     server_call_factories_and_concurrencies->emplace_back(
         std::move(mark_job_finished_call_factory), 1);
   }
 
  private:
   /// The grpc async service object.
-  JobInfoGcsService::AsyncService service_;
+  JobInfoAccessService::AsyncService service_;
   /// The service handler that actually handle the requests.
-  JobInfoHandler &service_handler_;
+  JobInfoAccessHandler &service_handler_;
 };
 
 }  // namespace rpc
 }  // namespace ray
-
-#endif
